+++ src/lib/config/Manufacturers.ts
@@ -4,7 +4,7 @@ import JSON5 from "json5";
 import path from "path";
 import { ZWaveError, ZWaveErrorCodes } from "../error/ZWaveError";
 import log from "../log";
-import { configDir } from "./utils";
+import { configDir, throwInvalidConfig } from "./utils";
 
 const configPath = path.join(configDir, "manufacturers.json");
 
@@ -20,10 +20,7 @@ async function loadManufacturers(): Promise<Record<string, string>> {
 		const fileContents = await readFile(configPath, "utf8");
 		return JSON5.parse(fileContents);
 	} catch (e) {
-		throw new ZWaveError(
-			"The config file is malformed!",
-			ZWaveErrorCodes.Config_Invalid,
-		);
+		throwInvalidConfig("manufacturers");
 	}
 }
 
@@ -43,10 +40,13 @@ export async function lookupManufacturer(
 				e instanceof ZWaveError &&
 				e.code === ZWaveErrorCodes.Config_Invalid
 			) {
-				log.driver.print(
-					`Could not load manufacturer config: ${e.message}`,
-					"error",
-				);
+				if (process.env.NODE_ENV !== "test") {
+					// FIXME: This call breaks when using jest.isolateModule()
+					log.driver.print(
+						`Could not load manufacturer config: ${e.message}`,
+						"error",
+					);
+				}
 				manufacturers = {};
 			} else {
 				// This is an unexpected error